% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd_graphs.R
\name{create_location_matrix}
\alias{create_location_matrix}
\title{Create Location Matrix for Peak Persistence Diagram}
\usage{
create_location_matrix(locs, labels, lambda_values)
}
\arguments{
\item{locs}{List of peak locations for each lambda value}

\item{labels}{List of peak labels for each lambda value}

\item{lambda_values}{Numeric vector of lambda values}
}
\value{
A matrix where rows correspond to lambda values and columns to peak labels
}
\description{
Create Location Matrix for Peak Persistence Diagram
}
\keyword{internal}
