% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd_graphs.R
\name{create_persistence_diagram_bc}
\alias{create_persistence_diagram_bc}
\title{Create Peak Persistence Diagram in Barchart Form}
\usage{
create_persistence_diagram_bc(intervals)
}
\arguments{
\item{intervals}{List where names are peak labels and values are vectors of lambda values
where the peak is significant}
}
\value{
A ggplot2 object representing the persistence diagram
}
\description{
Create Peak Persistence Diagram in Barchart Form
}
\examples{
\dontrun{
# Assuming 'peak_intervals' is a list of significant lambda values for each peak
persistence_plot <- create_persistence_diagram_bc(peak_intervals)
print(persistence_plot)
}
}
