% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd_util.R
\name{find_optimal_lambda}
\alias{find_optimal_lambda}
\title{Find Optimal Lambda Index}
\usage{
find_optimal_lambda(significant_peaks, persistent_labels)
}
\arguments{
\item{significant_peaks}{List of significant peak labels for each lambda}

\item{persistent_labels}{Numeric vector of persistent peak labels}
}
\value{
Index of optimal lambda value
}
\description{
Find Optimal Lambda Index
}
