% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd_util.R
\name{get_lambda}
\alias{get_lambda}
\title{Extract Lambda Value from Column Name}
\usage{
get_lambda(col_name)
}
\arguments{
\item{col_name}{String column name containing a lambda value}
}
\value{
Numeric lambda value
}
\description{
Extract Lambda Value from Column Name
}
