% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd_util.R
\name{get_peak_intervals}
\alias{get_peak_intervals}
\title{Get Peak Intervals}
\usage{
get_peak_intervals(significant_peaks)
}
\arguments{
\item{significant_peaks}{List of significant peak labels for each lambda}
}
\value{
List where names are peak labels and values are lambda values where the peak is significant
}
\description{
Get Peak Intervals
}
