% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_persistent_peaks.R
\name{get_persistent_peaks_clustering}
\alias{get_persistent_peaks_clustering}
\title{Find Persistent Peaks Using Clustering Method}
\usage{
get_persistent_peaks_clustering(peak_lists)
}
\arguments{
\item{peak_lists}{List of significant peaks for each lambda value}
}
\value{
Numeric vector of persistent peak labels
}
\description{
Find Persistent Peaks Using Clustering Method
}
