% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_persistent_peaks.R
\name{get_persistent_peaks_threshold}
\alias{get_persistent_peaks_threshold}
\title{Find Persistent Peaks Using a Threshold Method}
\usage{
get_persistent_peaks_threshold(label_intervals, ratio_threshold)
}
\arguments{
\item{label_intervals}{List of peak label intervals}

\item{ratio_threshold}{Numeric threshold between 0 and 1 for determining persistence}
}
\value{
Numeric vector of persistent peak labels
}
\description{
Find Persistent Peaks Using a Threshold Method
}
