% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd_util.R
\name{get_significant_peaks}
\alias{get_significant_peaks}
\title{Get Significant Peaks}
\usage{
get_significant_peaks(curvatures, labels, tau)
}
\arguments{
\item{curvatures}{Numeric vector of peak curvatures}

\item{labels}{Numeric vector of peak labels}

\item{tau}{Significance threshold for curvatures}
}
\value{
Numeric vector of significant peak labels
}
\description{
Get Significant Peaks
}
