% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_function.R
\name{reconstruct_function}
\alias{reconstruct_function}
\title{Reconstruct Function from Parameters}
\usage{
reconstruct_function(param, t_grid, idx, construct_data, smooth_object)
}
\arguments{
\item{param}{Numeric vector of parameters (basis coefficients and heights)}

\item{t_grid}{Numeric vector of time grid points}

\item{idx}{Numeric vector of indices for extrema points}

\item{construct_data}{Data frame containing time grid data}

\item{smooth_object}{Smooth object from mgcv}
}
\value{
A functional data object of class tf
}
\description{
Reconstructs a functional data object from optimization parameters, combining
basis coefficients and height values at extrema.
}
