% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_constrained_estimation.R
\name{shape_constrained_estimation}
\alias{shape_constrained_estimation}
\title{Shape Constrained Function Estimation}
\usage{
shape_constrained_estimation(
  curve_data,
  peak_locs,
  valley_locs,
  significant_peaks,
  peak_labels,
  mean_function,
  basis_dim = 8,
  t_grid,
  rho = 1e-09,
  basis_type = "cr"
)
}
\arguments{
\item{curve_data}{Functional data object of class tf}

\item{peak_locs}{Numeric vector of peak locations}

\item{valley_locs}{Numeric vector of valley locations}

\item{significant_peaks}{Numeric vector of significant peak indices}

\item{peak_labels}{Numeric vector of peak labels}

\item{mean_function}{Functional data object representing the mean function}

\item{basis_dim}{Integer specifying the basis dimension. Default is 8.}

\item{t_grid}{Numeric vector representing the time grid}

\item{rho}{Numeric smoothness penalty parameter}

\item{basis_type}{Character specifying basis type (default: "cr" for cubic regression splines)}
}
\value{
A functional data object of class tf representing the estimated function
}
\description{
Performs function estimation constrained by peak and valley locations
}
