% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize_warping_functions}
\alias{visualize_warping_functions}
\title{Visualize Warping Functions}
\usage{
visualize_warping_functions(warp_functions)
}
\arguments{
\item{warp_functions}{A tidyfun data object containing warping functions.}
}
\value{
A ggplot object displaying the warping functions.
}
\description{
Creates a plot of time warping functions to show how they distort the
time axis. The identity function (y = x) is shown as a dotted red line
for reference.
}
